Switched to C with SDL2
I found doing some thing as basic as changing the font using the Python libraries are impossible.
and the error handling is non-existent. so I will never know is the font imported or not.
so this repository will be called "Bevy-C-GUI" from now on. I setup the memory leak and dangling pointer defense with cppcheck and intel inspector.
I use C because I don't want people littering on my project with namespaces or OOP stuff.