Switched to Odin Raylib
Because it's way easier to read and write. It has good VSCodium support. It has "go to definition" and autocomplete feature. Which makes it easier to read Raylib source code without try to hunt where is the function in the file. And what is the function gonna do? What does it receive and return? I tried out Julia, Nim, Python, Nelua, Rust, C/C++. All of them are pain to use. It's either error prone with no error handling, or the code is so verbose that it's not even worth to use it. Or nasty memory leaks that will never get cleaned up. Raylib in Odin has no memory leak, except the memory leaks that comes from the dynamic-link library for windows kernel. But kernel leaks is not my concern.