Validation layers 50% | C++ prototype
i decided to covert the C++ code into DLL instead of trying to write it in Odin. then i will use Odin to control the C++ code. because Khronos might decide to break the API in the next day. and they use macros on debugCallback(), Odin doesn't support it. and the Vulkan bindings is a pain to maintain due to its size.