Add some basic OpenGL maths
I reverse engineered the GLM library. The library was unreadable, I found the code in some place are hand unrolled loop. I hate that. So I rolled the loop back in my code.
The rdrect still doesn't render. I am sure it's one of the step did wrong in OpenGL. I am still trying to find the bug. Which is impossible.
I am planning a header reorganize.
The code will written in ASM style or procedural programming instead of pure functions for consistent reason and avoiding malloc at the same time. It's awkward to pass the function allocated heap around in C.