Added LTO and AVX2 optimization
Now it will run like a train. And I am not planning to support older CPUs without AVX2. Which is before 2011.
But I am relying on auto-vectorization. Because I don't want to take the risk and waste my time. So it's possible to just remove the AVX2 compiling option and get it to work on your system.